feat: 특정 테마에 대한 예약이 존재하는데, 그 테마를 삭제하려 할 때 400 코드를 반환하는 기능 추가

+++ src/main/java/roomescape/service/ThemeService.java
@@ -2,19 +2,23 @@ package roomescape.service;
 
 import java.util.List;
 import org.springframework.stereotype.Service;
+import roomescape.dao.ReservationDao;
 import roomescape.dao.ThemeDao;
 import roomescape.domain.Theme;
 import roomescape.dto.request.ThemeRequest;
 import roomescape.dto.response.ThemeResponse;
 import roomescape.exception.ResourceNotExistException;
+import roomescape.exception.ThemeConstraintException;
 
 @Service
 public class ThemeService {
 
     private final ThemeDao themeDao;
+    private final ReservationDao reservationDao;
 
-    public ThemeService(ThemeDao themeDao) {
+    public ThemeService(ThemeDao themeDao, ReservationDao reservationDao) {
         this.themeDao = themeDao;
+        this.reservationDao = reservationDao;
     }
 
     public List<ThemeResponse> getAll() {
@@ -34,6 +38,11 @@ public class ThemeService {
     }
 
     public void deleteById(Long id) {
+        if (reservationDao.getCountByThemeId(id) != 0) {
+            // TODO: dao에서 던지는 예외와 service에서 던지는 예외가 과연 같아도 되는가
+            throw new ThemeConstraintException();
+        }
+
         int count = themeDao.deleteById(id);
         if (count == 0) {
             throw new ResourceNotExistException();
